+++ saver.h
@@ -10,7 +10,7 @@ public:
     Saver(const char *outFileName);
     virtual ~Saver();
 
-    virtual void save(float time) = 0;
+    virtual void save(double time) = 0;
 
 protected:
     std::ofstream _outFile;
